pyproject.toml
[project]
name = "mz_raport"
version = "0.1.0"
description = "This project is part of a recruitment task for the Ministry of Health."
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
authors = [
    { name = "Hubert Szewczyk", email = "hubert.szewczyk95@gmail.com" }
]

# Main dependencies
dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "python-dotenv>=0.19.0",
    "matplotlib>=3.4.0",
    "seaborn>=0.11.0",
]

# Optional dependencies
[project.optional-dependencies]
dev = [
    "pytest>=6.2.5",
    "black>=21.5b2",
    "flake8>=3.9.2",
    "isort>=5.9.1",
]
jupyter = [
    "jupyter",
    "ipykernel",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

# Tool configurations
[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]

[tool.flake8]
max-line-length = 88
extend-ignore = "E203"
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    "*.egg-info",
]
